+++ src/Model/SealSign.php
@@ -94,6 +94,7 @@ class SealSign implements ModelInterface, ArrayAccess
         'note' => '?string',
         'note_metadata' => '\DocuSign\eSign\Model\PropertyMetadata',
         'phone_authentication' => '\DocuSign\eSign\Model\RecipientPhoneAuthentication',
+        'proof_file' => '\DocuSign\eSign\Model\RecipientProofFile',
         'recipient_attachments' => '\DocuSign\eSign\Model\RecipientAttachment[]',
         'recipient_authentication_status' => '\DocuSign\eSign\Model\AuthenticationStatus',
         'recipient_feature_metadata' => '\DocuSign\eSign\Model\FeatureAvailableMetadata[]',
@@ -162,6 +163,7 @@ class SealSign implements ModelInterface, ArrayAccess
         'note' => null,
         'note_metadata' => null,
         'phone_authentication' => null,
+        'proof_file' => null,
         'recipient_attachments' => null,
         'recipient_authentication_status' => null,
         'recipient_feature_metadata' => null,
@@ -251,6 +253,7 @@ class SealSign implements ModelInterface, ArrayAccess
         'note' => 'note',
         'note_metadata' => 'noteMetadata',
         'phone_authentication' => 'phoneAuthentication',
+        'proof_file' => 'proofFile',
         'recipient_attachments' => 'recipientAttachments',
         'recipient_authentication_status' => 'recipientAuthenticationStatus',
         'recipient_feature_metadata' => 'recipientFeatureMetadata',
@@ -319,6 +322,7 @@ class SealSign implements ModelInterface, ArrayAccess
         'note' => 'setNote',
         'note_metadata' => 'setNoteMetadata',
         'phone_authentication' => 'setPhoneAuthentication',
+        'proof_file' => 'setProofFile',
         'recipient_attachments' => 'setRecipientAttachments',
         'recipient_authentication_status' => 'setRecipientAuthenticationStatus',
         'recipient_feature_metadata' => 'setRecipientFeatureMetadata',
@@ -387,6 +391,7 @@ class SealSign implements ModelInterface, ArrayAccess
         'note' => 'getNote',
         'note_metadata' => 'getNoteMetadata',
         'phone_authentication' => 'getPhoneAuthentication',
+        'proof_file' => 'getProofFile',
         'recipient_attachments' => 'getRecipientAttachments',
         'recipient_authentication_status' => 'getRecipientAuthenticationStatus',
         'recipient_feature_metadata' => 'getRecipientFeatureMetadata',
@@ -509,6 +514,7 @@ class SealSign implements ModelInterface, ArrayAccess
         $this->container['note'] = isset($data['note']) ? $data['note'] : null;
         $this->container['note_metadata'] = isset($data['note_metadata']) ? $data['note_metadata'] : null;
         $this->container['phone_authentication'] = isset($data['phone_authentication']) ? $data['phone_authentication'] : null;
+        $this->container['proof_file'] = isset($data['proof_file']) ? $data['proof_file'] : null;
         $this->container['recipient_attachments'] = isset($data['recipient_attachments']) ? $data['recipient_attachments'] : null;
         $this->container['recipient_authentication_status'] = isset($data['recipient_authentication_status']) ? $data['recipient_authentication_status'] : null;
         $this->container['recipient_feature_metadata'] = isset($data['recipient_feature_metadata']) ? $data['recipient_feature_metadata'] : null;
@@ -1373,6 +1379,30 @@ class SealSign implements ModelInterface, ArrayAccess
     public function setPhoneAuthentication($phone_authentication)
     {
         $this->container['phone_authentication'] = $phone_authentication;
+
+        return $this;
+    }
+
+    /**
+     * Gets proof_file
+     *
+     * @return \DocuSign\eSign\Model\RecipientProofFile
+     */
+    public function getProofFile()
+    {
+        return $this->container['proof_file'];
+    }
+
+    /**
+     * Sets proof_file
+     *
+     * @param \DocuSign\eSign\Model\RecipientProofFile $proof_file 
+     *
+     * @return $this
+     */
+    public function setProofFile($proof_file)
+    {
+        $this->container['proof_file'] = $proof_file;
 
         return $this;
     }