+++ CHANGELOG.md
@@ -3,9 +3,19 @@ All notable changes to this project will be documented in this file.
 
 See [DocuSign Support Center](https://support.docusign.com/en/releasenotes/) for Product Release Notes.
 
-## [V8.2.0] - eSignature API v2.1-25.1.00.02 - 2025-05-24 
+## [v8.4.0] - eSignature API v2.1-25.2.02.00 - 2025-08-11
 ### Changed
-- Added support for version v2.1-25.1.00.02 of the DocuSign ESignature API.
+- Added support for version v2.1-25.2.02.00 of the DocuSign ESignature API.
+- Updated the SDK release version.
+
+## [v8.3.0] - eSignature API v2.1-25.2.01.00 - 2025-07-08
+### Changed
+- Added support for version v2.1-25.2.01.00 of the DocuSign ESignature API.
+- Updated the SDK release version.
+
+## [v8.2.0] - eSignature API v2.1-25.2.00.00 - 2025-05-21
+### Changed
+- Added support for version v2.1-25.2.00.00 of the DocuSign ESignature API.
 - Updated the SDK release version.
 - Modernized constructor classes in Api, Client and Model files (including Configuration.php file) by adding `?array $data = null` to address deprecated warnings in PHP.
 - Added return types (`bool`, `mixed`, and `void`) to several functions in model files to improve type safety and compatibility with modern PHP versions.